feat: add key, keyboard and textfield
This commit introduces the key, keyboard and textfield element.
The large commit is needed as all three components are bound together
and need each other. Key is part of keyboard and a keyboard is needed
when there is a textfield.
The textfield is single line and touch-support only (like original).